% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battle.R
\encoding{UTF-8}
\name{battle}
\alias{battle}
\title{Run Pokémon Battle}
\usage{
battle(team_1, team_2, cpu = TRUE, generation = 8L)
}
\arguments{
\item{team_1}{Player 1's Pokémon team}

\item{team_2}{Player 2's Pokémon team}

\item{cpu}{Logical, is player 2 a CPU? Default set to \code{TRUE}}

\item{generation}{If teams are random, then the generation the Pokémon will be selected from.
Default is Gen 8.}
}
\description{
Create two teams of Pokémon and play out a match to see who is the world's number one Pokémon master!
}
\details{
For the sake of generation consistency, both teams must be pre-defined or both teams are random.
To create a random team use \code{\link{team}(generation = 1L)}.

In the initial version, the CPU will use a random move. AI will improve in future releases.
}
