% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_chance.R
\name{calculate_critical_chance}
\alias{calculate_critical_chance}
\alias{calculate_critical_chance_v1}
\alias{calculate_critical_chance_v2}
\title{Calculate Critical Hit Chance}
\usage{
calculate_critical_chance(move, ..., generation = 1L)

calculate_critical_chance_v1(move, speed, high_crit = 0L)

calculate_critical_chance_v2(move)
}
\arguments{
\item{move}{The name of the move that is being used}

\item{...}{Parameters that will be sent to calculations dependent on the generation}

\item{generation}{The generation that the battle is based on}

\item{speed}{The base speed of the \code{\link{Pokemon}} that has used the attack}

\item{high_crit}{Logical: has the \code{\link{Pokemon}} used a move that has increased
its critical hit chance?}
}
\description{
This calculates the probability that a critical hit will land.
}
\details{
Generation 1 does include the mechanics within the original games. For example, the
chance for a critical hit from standard moves \strong{reduces} after a move like swords
dance has been used.
}
\seealso{
https://bulbapedia.bulbagarden.net/wiki/Critical_hit
}
