% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage.R
\name{calculate_damage}
\alias{calculate_damage}
\title{Calculate Attack Damage}
\usage{
calculate_damage(
  power,
  p1_attack,
  p2_defense,
  p1_stab,
  p2_types,
  ...,
  level = 50L,
  critical = FALSE,
  generation = 1L
)
}
\description{
Calculate Attack Damage
}
