% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp.R
\encoding{UTF-8}
\name{calculate_hp}
\alias{calculate_hp}
\alias{calculate_hp_range}
\title{Calculating Pokémon Hit Points}
\usage{
calculate_hp(base_hp, level = 50, iv = 0, ev = 0, generation = 1L)

calculate_hp_range(base_hp, level = 50, generation = 1L)
}
\arguments{
\item{base_hp}{The base hit point stat a Pokémon has}

\item{level}{The level of the Pokémon}

\item{iv}{Individual Value. A random number given to a Pokémon
between 1-15 (Generation 1-2)  or 1-31 (Generation 3+)}

\item{ev}{Effort Value.}

\item{generation}{Generation to look at. Calculations differ in generations
1 and 2 to subsequent generations}
}
\description{
Find the HP of a Pokémon given its base HP stat, along with other
information such as its level, IVs and EVs, and the generation.

\code{calculate_hp_range} gives a potential range of the HP stat, though
this is assuming that no EVs have been added to the Pokémon.
}
\examples{
# TODO with real pokemon stats

}
\seealso{
\code{\link{calculate_stat}}
}
