% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\encoding{UTF-8}
\name{calculate_stat}
\alias{calculate_stat}
\alias{calculate_stat_range}
\title{Calculating Pokémon Stats}
\usage{
calculate_stat(
  base_stat,
  stat_name,
  nature,
  level = 50,
  iv = 0,
  ev = 0,
  generation = 1L
)

calculate_stat_range(base_stat, stat_name, nature, level = 50, generation = 1L)
}
\arguments{
\item{base_stat}{The stat a Pokémon has}

\item{stat_name}{The name of the stat i.e attack. Ignored for earlier
generations.}

\item{nature}{Nature of the Pokémon. Ignored for earlier generations.}

\item{level}{The level of the Pokémon}

\item{iv}{Individual Value. A random number given to a Pokémon
between 1-15 (Generation 1-2)  or 1-31 (Generation 3+)}

\item{ev}{Effort Value.}

\item{generation}{Generation to look at. Calculations differ in generations
1 and 2 to subsequent generations}
}
\description{
Find the attack, defense, special attack, special defense and speed
of a Pokémon given its base stat value, along with other
information such as its level, IVs and EVs, and the generation.

\code{calculate_hp_range} gives a potential range of the HP stat, though
this is assuming that no EVs have been added to the Pokémon.
}
\examples{
# TODO with real Pokémon stats

}
\seealso{
\code{\link{calculate_hp}}
}
