% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_multiplier.R
\name{calculate_stat_multiplier}
\alias{calculate_stat_multiplier}
\title{Calculate Stat Multiplier}
\usage{
calculate_stat_multiplier(stat, stage = 0L, generation = 1L)
}
\arguments{
\item{stat}{Name of the stat}

\item{stage}{Number of stages the stat has increased/decreased}

\item{generation}{Generation of the batlle}
}
\description{
Given a stat and the amount increased/decreased, a multiplier can be applied to that stat
}
\seealso{
\url{https://bulbapedia.bulbagarden.net/wiki/Stat_modifier#In-battle_modification}
}
