% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage_utils.R
\name{include_stab_multiplier}
\alias{include_stab_multiplier}
\title{STAB Multiplier}
\usage{
include_stab_multiplier(move_type, pokemon_types)
}
\arguments{
\item{move_type}{The ID of type of the move used}

\item{pokemon_types}{A vector of the IDs of types of the attacking Pokémon}
}
\description{
Finding out whether the type of the moved used matches the attacking Pokémon type.
If so then a multiplier of 1.5 is given, otherwise it is 1.
}
