% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\encoding{UTF-8}
\name{learn_moves}
\alias{learn_moves}
\alias{choose_moves}
\title{Learn Pokémon Moves}
\usage{
learn_moves(moves, level = 50L, generation = 1L, n = 4L)

choose_moves(moves, level = 50L, generation = 1L, n = 4L)
}
\arguments{
\item{moves}{A list of moves extracted from \code{pokeapi::get_pokemon}}

\item{level}{The level of the Pokémon. Used to work out which
moves are viable}

\item{generation}{The generation to look at. Used to work out
which moves are viable}

\item{n}{The maximum number of moves learnt by the Pokémon. Must be
between 1 and 4 (default is 4).}
}
\description{
Learn Pokémon Moves
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pikachu <- pokeapi::get_pokemon(25)
pikachu_moves <- learn_moves(pikachu$moves)
\dontshow{\}) # examplesIf}
}
