% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\encoding{UTF-8}
\name{Move}
\alias{Move}
\title{Pokémon Move}
\description{
R6 object containing required information about a Pokémon move.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-move-new}{\code{Move$new()}}
\item \href{#method-move-get_stat}{\code{Move$get_stat()}}
\item \href{#method-move-clone}{\code{Move$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move-new"></a>}}
\if{latex}{\out{\hypertarget{method-move-new}{}}}
\subsection{Method \code{new()}}{
Create a Move
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Move$new(move_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{move_name}}{The hyphen cased name of the move e.g. \code{"razor-leaf"}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Move
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move-get_stat"></a>}}
\if{latex}{\out{\hypertarget{method-move-get_stat}{}}}
\subsection{Method \code{get_stat()}}{
Get the stat of the move
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Move$get_stat(stat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stat}}{The private field of the move}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move-clone"></a>}}
\if{latex}{\out{\hypertarget{method-move-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Move$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
