% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\encoding{UTF-8}
\name{new_pokemon}
\alias{new_pokemon}
\title{Create a New Pokémon}
\usage{
new_pokemon(
  api_data = NULL,
  level = 50L,
  generation = 8,
  nature = NULL,
  language = "en"
)
}
\description{
Create a New Pokémon
}
