% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\encoding{UTF-8}
\name{pokemon}
\alias{pokemon}
\title{Create a New Pokémon}
\usage{
pokemon(
  api_data = NULL,
  pokemon = NULL,
  level = 50L,
  generation = 8L,
  nature = NULL,
  language = "en"
)
}
\arguments{
\item{api_data}{Either data pulled from \code{pokeapi::get_pokemon}, or
\code{NULL} to select the data of a random Pokémon}

\item{pokemon}{Name of the Pokémon to create}

\item{level}{The level the Pokémon should be}

\item{generation}{The generation that the Pokémon comes from}

\item{nature}{Either a specified nature, or \code{NULL} for a random nature.
\strong{NB} Nature is only required from generation 3 onwards}

\item{language}{ISO-2 character of the language to pull the name of
the Pokémon}
}
\value{
A Pokémon.
}
\description{
R7 object containing required information about a Pokémon.
}
