% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battle.R
\encoding{UTF-8}
\name{PokemonBattle}
\alias{PokemonBattle}
\title{Run Pokémon Battle}
\description{
Create two teams of Pokémon and play out a match to see who is the world's number one Pokémon master!

To create a random team use \code{\link{PokemonTeam}$new(generation = 1L)}.

In the initial version, the CPU will use a random move. AI will improve in future releases.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pokemon_battle-new}{\code{PokemonBattle$new()}}
\item \href{#method-pokemon_battle-start}{\code{PokemonBattle$start()}}
\item \href{#method-pokemon_battle-status}{\code{PokemonBattle$status()}}
\item \href{#method-pokemon_battle-attack}{\code{PokemonBattle$attack()}}
\item \href{#method-pokemon_battle-attack_p2}{\code{PokemonBattle$attack_p2()}}
\item \href{#method-pokemon_battle-switch}{\code{PokemonBattle$switch()}}
\item \href{#method-pokemon_battle-switch_p2}{\code{PokemonBattle$switch_p2()}}
\item \href{#method-pokemon_battle-clone}{\code{PokemonBattle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_battle-new"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_battle-new}{}}}
\subsection{Method \code{new()}}{
Commence a Pokémon battle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$new(
  team_1,
  team_2,
  player_2_cpu = TRUE,
  level = 50L,
  generation = 1L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team_1}}{Player 1's Pokémon team}

\item{\code{team_2}}{Player 2's Pokémon team}

\item{\code{player_2_cpu}}{Logical, is player 2 a CPU? Default set to \code{TRUE}}

\item{\code{level}}{If random teams, the level the Pokémon should be}

\item{\code{generation}}{If teams are random, then the generation the Pokémon will be selected from.
Default is Gen 8.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A state where a Pokemon battle can commence
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_battle-start"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_battle-start}{}}}
\subsection{Method \code{start()}}{
A helper to start the local console version of a battle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_battle-status"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_battle-status}{}}}
\subsection{Method \code{status()}}{
Show current status of Pokémon Battle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_battle-attack"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_battle-attack}{}}}
\subsection{Method \code{attack()}}{
Select Pokémon attack
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$attack()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_battle-attack_p2"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_battle-attack_p2}{}}}
\subsection{Method \code{attack_p2()}}{
Select Pokémon attack (player 2)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$attack_p2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_battle-switch"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_battle-switch}{}}}
\subsection{Method \code{switch()}}{
Select active Pokémon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$switch(new_active, curr_active = private$active_1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_active}}{Team position of the Pokémon to switch in}

\item{\code{curr_active}}{Team position of the Pokémon to switch out. Defaults to current active}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_battle-switch_p2"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_battle-switch_p2}{}}}
\subsection{Method \code{switch_p2()}}{
Select active Pokémon (player 2)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$switch_p2(new_active, curr_active = private$active_2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_active}}{Team position of the Pokémon to switch in}

\item{\code{curr_active}}{Team position of the Pokémon to switch out. Defaults to current active}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_battle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_battle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
