% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_print.R
\encoding{UTF-8}
\name{showStatus}
\alias{showStatus}
\title{Print Current Pokémon Status}
\usage{
showStatus(x, simple = FALSE)
}
\description{
A way to show the current stats and status of a selected Pokémon.

Includes:
\itemize{
\item Pokémon base stats and health
\item Pokemon move set and PP
}
}
