% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\encoding{UTF-8}
\name{team}
\alias{team}
\title{Pokémon Team}
\usage{
team(
  pokemon_1,
  pokemon_2,
  pokemon_3,
  pokemon_4,
  pokemon_5,
  pokemon_6,
  random = TRUE,
  generation = 8L
)
}
\arguments{
\item{pokemon_1, pokemon_2, pokemon_3, pokemon_4, pokemon_5, pokemon_6}{Pre-created Pokémon}

\item{random}{Logical, should the team be randomised? Default set to \code{TRUE}}

\item{generation}{The generation that the Pokémon comes from}
}
\value{
A Team of 6 Pokémon.
}
\description{
R7 object containing required information about a Pokémon.
}
