% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack.R
\name{use_attack}
\alias{use_attack}
\title{Use Attack}
\usage{
use_attack(move, attacker, defender, battle, generation = 1L)
}
\arguments{
\item{move}{Name of the move that the \code{atttacker} is using}

\item{attacker}{The attacking \code{\link{Pokemon}}}

\item{defender}{The defending \code{\link{Pokemon}}}

\item{battle}{The \code{\link{PokemonBattle}} in place. Used to know}

\item{generation}{The generation that the battle is happening in}
}
\value{
Nothing returned, however the \code{attacker}, \code{defender} and \code{battle}
are all updated to the effects of the \code{move} used.
}
\description{
Calculate the effect of the attack
}
