% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_chance.R
\name{calculate_critical_chance}
\alias{calculate_critical_chance}
\title{Calculate Critical Hit Chance}
\usage{
calculate_critical_chance(move, ..., generation = 1L)
}
\description{
Calculate Critical Hit Chance
}
\seealso{
https://bulbapedia.bulbagarden.net/wiki/Critical_hit
}
