% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\encoding{UTF-8}
\name{Move}
\alias{Move}
\title{Pokémon Move}
\description{
R6 object containing required information about a Pokémon move.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-move-new}{\code{Move$new()}}
\item \href{#method-move-get_pp_status}{\code{Move$get_pp_status()}}
\item \href{#method-move-get_stat}{\code{Move$get_stat()}}
\item \href{#method-move-can_use_move}{\code{Move$can_use_move()}}
\item \href{#method-move-use_move}{\code{Move$use_move()}}
\item \href{#method-move-clone}{\code{Move$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move-new"></a>}}
\if{latex}{\out{\hypertarget{method-move-new}{}}}
\subsection{Method \code{new()}}{
Create a Move
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Move$new(move_name, generation = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{move_name}}{The hyphen cased name of the move e.g. \code{"razor-leaf"}}

\item{\code{generation}}{The generation that the Pokémon comes from}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Move
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move-get_pp_status"></a>}}
\if{latex}{\out{\hypertarget{method-move-get_pp_status}{}}}
\subsection{Method \code{get_pp_status()}}{
Get current PP of the move
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Move$get_pp_status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move-get_stat"></a>}}
\if{latex}{\out{\hypertarget{method-move-get_stat}{}}}
\subsection{Method \code{get_stat()}}{
Get the stat of the move
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Move$get_stat(stat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stat}}{The private field of the move}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move-can_use_move"></a>}}
\if{latex}{\out{\hypertarget{method-move-can_use_move}{}}}
\subsection{Method \code{can_use_move()}}{
Checks the ability to use the move
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Move$can_use_move()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical as to whether or not the move has been used or not
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move-use_move"></a>}}
\if{latex}{\out{\hypertarget{method-move-use_move}{}}}
\subsection{Method \code{use_move()}}{
Checks the ability to use the move
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Move$use_move(attacker, defender, battle, generation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attacker}}{The attacking \code{\link{Pokemon}}}

\item{\code{defender}}{The defending \code{\link{Pokemon}}}

\item{\code{battle}}{The \code{\link{PokemonBattle}} in place. Used to know}

\item{\code{generation}}{The generation that the battle is happening in}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical as to whether or not the move has been used or not
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move-clone"></a>}}
\if{latex}{\out{\hypertarget{method-move-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Move$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
