% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\encoding{UTF-8}
\name{Pokemon}
\alias{Pokemon}
\title{Create a New Pokémon}
\description{
R6 object containing required information about a Pokémon.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pokemon-new}{\code{Pokemon$new()}}
\item \href{#method-pokemon-status}{\code{Pokemon$status()}}
\item \href{#method-pokemon-use_move}{\code{Pokemon$use_move()}}
\item \href{#method-pokemon-get_stat}{\code{Pokemon$get_stat()}}
\item \href{#method-pokemon-get_moves}{\code{Pokemon$get_moves()}}
\item \href{#method-pokemon-change_move}{\code{Pokemon$change_move()}}
\item \href{#method-pokemon-get_crit_chance}{\code{Pokemon$get_crit_chance()}}
\item \href{#method-pokemon-clone}{\code{Pokemon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon-new"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon-new}{}}}
\subsection{Method \code{new()}}{
Create a Pokémon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$new(
  pokemon = NULL,
  level = 50L,
  generation = 1L,
  nature = NULL,
  api_data = NULL,
  language = "en"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pokemon}}{Name of the Pokémon to create}

\item{\code{level}}{The level the Pokémon should be}

\item{\code{generation}}{The generation that the Pokémon comes from}

\item{\code{nature}}{Either a specified nature, or \code{NULL} for a random nature.
\strong{NB} Nature is only required from generation 3 onwards}

\item{\code{api_data}}{Either data pulled from \code{pokeapi::get_pokemon}, or
\code{NULL} to select the data of a random Pokémon}

\item{\code{language}}{ISO-2 character of the language to pull the name of
the Pokémon}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Pokémon
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon-status"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon-status}{}}}
\subsection{Method \code{status()}}{
Show current status of the Pokémon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$status(simple = FALSE, console = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simple}}{Logical, do you just want the simple status (name + HP) printed?}

\item{\code{console}}{Logical, should the status be printed or returned as a string?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon-use_move"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon-use_move}{}}}
\subsection{Method \code{use_move()}}{
Use an attack of the Pokémon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$use_move(move, battle)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{move}}{The name of the move that is being used}

\item{\code{battle}}{A \code{\link{PokemonBattle}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon-get_stat"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon-get_stat}{}}}
\subsection{Method \code{get_stat()}}{
Get the stat of the Pokémon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$get_stat(stat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stat}}{The private field of the Pokémon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon-get_moves"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon-get_moves}{}}}
\subsection{Method \code{get_moves()}}{
Get the moveset of the Pokémon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$get_moves()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon-change_move"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon-change_move}{}}}
\subsection{Method \code{change_move()}}{
Change one of the moves of the Pokémon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$change_move(new_move = NULL, replace_move = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_move}}{Name of the new move}

\item{\code{replace_move}}{Either the name of the move to replace, or the position of the move}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon-get_crit_chance"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon-get_crit_chance}{}}}
\subsection{Method \code{get_crit_chance()}}{
Get the critical hit chance for a move used by the Pokémon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$get_crit_chance(move)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{move}}{Name of the move used by the Pokémon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pokemon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
