% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battle.R
\encoding{UTF-8}
\name{PokemonBattle}
\alias{PokemonBattle}
\title{Run Pokémon Battle}
\description{
Create two teams of Pokémon and play out a match to see who is the world's number one Pokémon master!
}
\details{
For the sake of generation consistency, both teams must be pre-defined or both teams are random.
To create a random team use \code{\link{team}(generation = 1L)}.

In the initial version, the CPU will use a random move. AI will improve in future releases.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-battle-new}{\code{PokemonBattle$new()}}
\item \href{#method-battle-status}{\code{PokemonBattle$status()}}
\item \href{#method-battle-clone}{\code{PokemonBattle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-battle-new"></a>}}
\if{latex}{\out{\hypertarget{method-battle-new}{}}}
\subsection{Method \code{new()}}{
Commence a Pokémon battle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$new(team_1, team_2, cpu = TRUE, generation = 8L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{team_1}}{Player 1's Pokémon team}

\item{\code{team_2}}{Player 2's Pokémon team}

\item{\code{cpu}}{Logical, is player 2 a CPU? Default set to \code{TRUE}}

\item{\code{generation}}{If teams are random, then the generation the Pokémon will be selected from.
Default is Gen 8.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-battle-status"></a>}}
\if{latex}{\out{\hypertarget{method-battle-status}{}}}
\subsection{Method \code{status()}}{
Show current status of Pokémon Battle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-battle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-battle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonBattle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
