% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\encoding{UTF-8}
\name{PokemonTeam}
\alias{PokemonTeam}
\title{Pokémon Team}
\description{
R6 object containing required information about your Pokémon team.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pokemon_team-new}{\code{PokemonTeam$new()}}
\item \href{#method-pokemon_team-status}{\code{PokemonTeam$status()}}
\item \href{#method-pokemon_team-clone}{\code{PokemonTeam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_team-new"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_team-new}{}}}
\subsection{Method \code{new()}}{
Create a new Pokémon team
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonTeam$new(
  pokemon_1,
  pokemon_2,
  pokemon_3,
  pokemon_4,
  pokemon_5,
  pokemon_6,
  random = FALSE,
  generation = 1L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pokemon_1, pokemon_2, pokemon_3, pokemon_4, pokemon_5, pokemon_6}}{Pre-created Pokémon}

\item{\code{random}}{Logical, should the team be randomised? Default set to \code{TRUE}}

\item{\code{generation}}{The generation that the Pokémon comes from}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Team of 6 Pokémon
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_team-status"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_team-status}{}}}
\subsection{Method \code{status()}}{
Show current status of Pokémon Team
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonTeam$status(simple = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simple}}{Logical, do you just want the simple status (name + HP) printed?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pokemon_team-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pokemon_team-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PokemonTeam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
