% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_print.R
\encoding{UTF-8}
\name{show_status}
\alias{show_status}
\title{Print Current Pokémon Status}
\usage{
show_status(x, simple = FALSE)
}
\arguments{
\item{x}{The private fields of a Pokémon}

\item{simple}{Logical, should a simplified print be used}
}
\description{
A way to show the current stats and status of a selected Pokémon.

Includes:
\itemize{
\item Pokémon base stats and health
\item Pokemon move set and PP
}
}
