% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_print.R
\encoding{UTF-8}
\name{status}
\alias{status}
\title{Print Current Pokémon Status}
\usage{
status(x)
}
\description{
A way to show the current stats and status of a selected Pokémon.

Includes:
\itemize{
\item Pokémon base stats and health
\item Pokemon move set and PP
}
}
